Binary equilibrium: Power market example
This is an illustrative example for the “multi-objective program
subject to a binary quasi-equilibrium (MOPBQE)” method using a power
market application with uplift payments.

The mathematical description is documented in chapter 4 of:
Daniel Huppmann and Sauleh Siddiqui.
"An exact solution method for binary equilibrium problems with
compensation and the power market uplift problem", DIW Discussion Paper
1475, 2015.